Moved REST API-specific browsermob exception classes to net.lightbody.bmp.exception in browsermob-rest
