+++ browsermob-rest/src/main/java/net/lightbody/bmp/exception/ProxyPortsExhaustedException.java
@@ -1,9 +1,9 @@
-package net.lightbody.bmp.proxy;
+package net.lightbody.bmp.exception;
 
 public class ProxyPortsExhaustedException extends RuntimeException {
-	private static final long serialVersionUID = 5365335130190989903L;
+    private static final long serialVersionUID = -6801448612785792233L;
 
-	public ProxyPortsExhaustedException() {
+    public ProxyPortsExhaustedException() {
 		super();
 	}
