+++ browsermob-rest/src/main/java/net/lightbody/bmp/proxy/bricks/ProxyResource.java
@@ -28,9 +28,9 @@ import javax.script.ScriptEngine;
 import javax.script.ScriptEngineManager;
 import javax.script.ScriptException;
 import net.lightbody.bmp.core.har.Har;
-import net.lightbody.bmp.proxy.ProxyExistsException;
+import net.lightbody.bmp.exception.ProxyExistsException;
 import net.lightbody.bmp.proxy.ProxyManager;
-import net.lightbody.bmp.proxy.ProxyPortsExhaustedException;
+import net.lightbody.bmp.exception.ProxyPortsExhaustedException;
 import net.lightbody.bmp.proxy.ProxyServer;
 import net.lightbody.bmp.proxy.http.BrowserMobHttpRequest;
 import net.lightbody.bmp.proxy.http.BrowserMobHttpResponse;
