+++ browsermob-rest/src/test/java/net/lightbody/bmp/proxy/ProxyPortAssignmentTest.java
@@ -4,6 +4,8 @@ import java.util.HashMap;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 
+import net.lightbody.bmp.exception.ProxyExistsException;
+import net.lightbody.bmp.exception.ProxyPortsExhaustedException;
 import net.lightbody.bmp.proxy.test.util.ProxyManagerTest;
 import org.junit.Test;
 
@@ -25,7 +27,7 @@ public class ProxyPortAssignmentTest extends ProxyManagerTest {
         try{
             proxyManager.create(new HashMap<String, String>());        
             fail();
-        }catch(ProxyPortsExhaustedException e){            
+        }catch(ProxyPortsExhaustedException e){
             proxyManager.delete(9093);
             p = proxyManager.create(new HashMap<String, String>());
             assertEquals(9093, p.getPort());
@@ -47,7 +49,7 @@ public class ProxyPortAssignmentTest extends ProxyManagerTest {
         try{            
             proxyManager.create(new HashMap<String, String>(), 9094);            
             fail();
-        }catch(ProxyExistsException e){            
+        }catch(ProxyExistsException e){
             assertEquals(9094, e.getPort());
             proxyManager.delete(9094);
         }
